requirements.txt
BlurWindow==1.2.1
PySide6==6.4.1
PySide6-Addons==6.4.1
PySide6-Essentials==6.4.1
pyxdg==0.28
shiboken6==6.4.1
